+++ Frieght.Api/Infrastructure/DataExtensions.cs
@@ -27,10 +27,10 @@ public static  class DataExtensions
     {
         // setitngs for docker container
         var connString = Environment.GetEnvironmentVariable("DefaultConnection"); // to retrieve connection from docker container environment variable   
+        
         // var connString = configuration.GetConnectionString("DefaultConnection"); // to retrieve connection from configuration file like appsettings.json
         // Console.WriteLine($"DefaultConnection: {connString}");
 
-        // var connString = configuration.GetConnectionString("DefaultConnection");
         services.AddDbContext<FrieghtDbContext>(options =>
             options.UseNpgsql(connString)) // Changed to UseNpgsql
             .AddScoped<ILoadRepository, LoadRepository>()