+++ Frieght.Api/Changelog.md
@@ -1,3 +1,4 @@
 # Pending
 
 * [Task 39]: Deploy frieight api and make it accessible
+* [Sprint 11]: Add getting bid by load and carrierId